requirements.txt
# Resume Automation System - Python Dependencies

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0

# Claude API
anthropic==0.7.1

# Document Generation
python-docx==0.8.11
reportlab==4.0.9
pypdf==3.17.1

# Database & ORM
sqlalchemy==2.0.23

# CLI & Utilities
click==8.1.7
python-dotenv==1.0.0
pandas==2.1.3

# Progress & Logging
tqdm==4.66.1
loguru==0.7.2

# Optional: Web UI (Phase 3)
# flask==3.0.0

# Development (optional)
# pytest==7.4.3
# black==23.12.0
# flake8==6.1.0
